wip: merging changes
