+++ rl/offpolicy/algorithms/sac.py
@@ -1,7 +1,7 @@
 import torch
 from catalyst.dl.utils import UtilsFactory
 import catalyst.rl.agents.agents as agents
-from catalyst.rl.algorithms.base import BaseAlgorithm, soft_update
+from catalyst.rl.offpolicy.algorithms.base import BaseAlgorithm, soft_update
 
 
 class SAC(BaseAlgorithm):
